+++ libfreerdp-rfx/rfx_dwt.h
@@ -23,7 +23,7 @@
 #include <freerdp/rfx.h>
 
 void
-rfx_dwt_2d_decode(RFX_CONTEXT * context, int * buffer, int subband_width);
+rfx_dwt_2d_decode(RFX_CONTEXT * context, short * buffer, int subband_width);
 
 #endif
